requirements.txt
requests>=2.25.0,<3.0.0
websockets>=10.0,<13.0.0
python-dotenv>=0.15.0,<2.0.0